+++ scripts/scala-build-7.sh
@@ -2,10 +2,17 @@
 
 set -e
 
+if [[ ! -v JAVA17 ]]; then
+  echo "Error: You must set the environment variable 'JAVA17'."
+  exit 1
+fi
+
 if [[ ! -v SCALA3 ]]; then
   echo "Error: You must set the environment variable 'SCALA3'."
   exit 1
 fi
+
+export JAVACMD=$JAVA17
 
 rm -rf out/lib/7.0.4
 mkdir -p out/lib/7.0.4
